Better 'multi' and reorganized code.
The ’multi’ version now uses subprocess instead of ‘&’ at the end of an
os.system command. This will be safer and should work  with more
operating systems. For some reason, this seems to make ‘multi’ slightly
slower, but not noticeably in most situations. Some other small changes
as well (print & debugmode stuff).